# docs/api/video-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ss-video",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/cvat/frames": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List frames pending annotation (for building a CVAT task)",
        "description": "Return frames that need annotation, ordered by al_score descending.\n\nal_tag: 'needs_annotation' | 'novel' | 'any' (both).\nlimit: max rows to return (1-5000, default 200).\n\nUse the returned frame_path list to create a CVAT task, then call\nPOST /admin/cvat/task to record the mapping for webhook resolution.",
        "operationId": "cvat_annotation_frames_admin_cvat_frames_get",
        "parameters": [
          {
            "name": "al_tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "needs_annotation",
              "title": "Al Tag"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvatFrameListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cvat/task": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Register a CVAT task \u2192 frame mapping (enables webhook resolution)",
        "description": "Store the mapping from a CVAT task ID to selfsuvis frame IDs.\n\nCall this after creating a CVAT task with frames from GET /admin/cvat/frames.\nWhen CVAT fires a webhook for job completion, this mapping is used to find\nand annotate the corresponding selfsuvis frames.\n\nIdempotent \u2014 duplicate (task_id, frame_id) pairs are silently ignored.",
        "operationId": "register_cvat_task_admin_cvat_task_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CvatTaskRegistration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Register Cvat Task Admin Cvat Task Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/webhook/cvat": {
      "post": {
        "tags": [
          "webhook"
        ],
        "summary": "Receive CVAT annotation webhook",
        "description": "Handle CVAT webhook events and update frame annotation status.\n\nTriggered by CVAT when a job or task reaches state=completed.\nMarks all frames mapped to that task as al_tag='annotated'.\n\nSecurity: verifies HMAC-SHA256 signature in X-Hook-Secret.\nCVAT_WEBHOOK_SECRET must be set; requests with a missing or incorrect\nsignature are rejected with 400. Configure the same secret in CVAT:\nWebhooks \u2192 Secret \u2192 <your CVAT_WEBHOOK_SECRET value>.\n\nConfigure target URL in CVAT: http://<api_host>:8000/webhook/cvat",
        "operationId": "cvat_webhook_webhook_cvat_post",
        "parameters": [
          {
            "name": "x-hook-secret",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Hook-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cvat Webhook Webhook Cvat Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CvatFrameItem": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id"
          },
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "frame_path": {
            "type": "string",
            "title": "Frame Path"
          },
          "al_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Al Score"
          },
          "al_tag": {
            "type": "string",
            "title": "Al Tag"
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "mission_id",
          "frame_path",
          "al_score",
          "al_tag"
        ],
        "title": "CvatFrameItem"
      },
      "CvatFrameListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/CvatFrameItem"
            },
            "type": "array",
            "title": "Frames"
          }
        },
        "type": "object",
        "required": [
          "total",
          "frames"
        ],
        "title": "CvatFrameListResponse"
      },
      "CvatTaskRegistration": {
        "properties": {
          "cvat_task_id": {
            "type": "integer",
            "title": "Cvat Task Id"
          },
          "frame_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Frame Ids"
          }
        },
        "type": "object",
        "required": [
          "cvat_task_id",
          "frame_ids"
        ],
        "title": "CvatTaskRegistration"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}